Configure react-native-paper and react-navigation theme and use `MaterialCommunityIcons` as default icons for react-native-paper (the latter does not work for the icons used in `app-navigator.tsx`)
