+++ app/app.tsx
@@ -11,11 +11,12 @@
  */
 import "react-native-gesture-handler"
 import React, { useEffect } from "react"
+import { useColorScheme } from "react-native"
 import { SafeAreaProvider, initialWindowMetrics } from "react-native-safe-area-context"
 import { initFonts } from "./theme/fonts" // expo
 import { useBackButtonHandler, AppNavigator, canExit, useNavigationPersistence } from "./navigators"
 import { ErrorBoundary } from "./components"
-import { configureFonts, DefaultTheme, Provider as PaperProvider } from "react-native-paper"
+import { Provider as PaperProvider } from "react-native-paper"
 
 // This puts screens in a native ViewController or Activity. If you want fully native
 // stack navigation, use `createNativeStackNavigator` in place of `createStackNavigator`:
@@ -26,40 +27,20 @@ import * as FileSystem from "expo-file-system"
 import { tracksDirectoryUri } from "./utils/file"
 import { TrackDownloadsClientProvider } from "./contexts/TrackDownloadsClientContext"
 import { TrackDownloadsClient } from "./clients/TrackDownloadsClient"
-import { Theme } from "react-native-paper/lib/typescript/types"
+import { MaterialCommunityIcons } from "@expo/vector-icons"
+import { navigatorTheme, theme } from "./theme"
 
 enableScreens()
 
 export const NAVIGATION_PERSISTENCE_KEY = "NAVIGATION_STATE"
-
-const theme: Theme = {
-  ...DefaultTheme,
-  dark: false,
-  roundness: 4,
-  colors: {
-    primary: "#6200ee",
-    accent: "#03dac4",
-    background: "#f6f6f6",
-    surface: "#ffffff",
-    error: "#B00020",
-    text: "#000000",
-    onSurface: "#000000",
-    disabled: "rgba(0, 0, 0, 0.26)",
-    placeholder: "rgba(0, 0, 0, 0.54)",
-    backdrop: "rgba(0, 0, 0, 0.5)",
-    notification: "#f50057",
-  },
-  fonts: configureFonts(),
-  animation: {
-    scale: 1.0,
-  },
-}
 
 /**
  * This is the root component of our app.
  */
 function App() {
   const [ready, setReady] = useState(false)
+
+  const colorScheme = useColorScheme()
 
   useBackButtonHandler(canExit)
   const {
@@ -82,11 +63,23 @@ function App() {
   return (
     <SafeAreaProvider initialMetrics={initialWindowMetrics}>
       <TrackDownloadsClientProvider client={new TrackDownloadsClient()}>
-        <PaperProvider theme={theme}>
+        <PaperProvider
+          theme={theme(colorScheme)}
+          settings={{
+            icon: ({ name, ...props }) => (
+              // TODO avoid unsafe cast of name
+              <MaterialCommunityIcons
+                name={name as React.ComponentProps<typeof MaterialCommunityIcons>["name"]}
+                {...props}
+              />
+            ),
+          }}
+        >
           <ErrorBoundary catchErrors={"always"}>
             <AppNavigator
               initialState={initialNavigationState}
               onStateChange={onNavigationStateChange}
+              theme={navigatorTheme(colorScheme)}
             />
           </ErrorBoundary>
         </PaperProvider>
